Moves rawDataToVlilleStation() to static method in VlilleStation model
